requirements.txt
discord.py>=2.0.0
asyncpg>=0.24.0
python-dotenv>=0.19.0